Track Load result dtypes from source values
Untyped ThreadData outputs do not carry a dtype for downstream
planning. Follow the Load source for dtype and the output descriptor
for shape so directly composed operations remain inferable.

+++ python/cuda_coop/tests/backends/numba_mlir/unit/test_data_movement.py
@@ -226,7 +226,7 @@ def test_common_direct_block_load_store_lowers_to_private_factories():
 
     def memory(source, destination):
         storage = coop.TempStorage()
-        output = coop.ThreadData(2, dtype=types.int32)
+        output = coop.ThreadData(2)
         loaded = coop.load(
             coop.this_block(),
             source,
